Organize editor pages and shared helpers
