+++ src/pages/editor/index.js
@@ -1,0 +1,4 @@
+export { default as EditorPage } from './EditorPage';
+export { default as V2EditorPage } from './V2EditorPage';
+export { default as EditorNewProjectPage } from './EditorNewProjectPage';
+export { default as EditorProjectsPage } from './EditorProjectsPage';
